Improved UI of the website
